move calculat_total_drilling_lentghs from economics to wellbores in prepartion for updating it
