+++ src/geophires_x/WellBores.py
@@ -1363,3 +1363,46 @@ class WellBores:
             self.PumpingPower.value = [0. if x < 0. else x for x in self.PumpingPower.value]
 
         model.logger.info(f'complete {self.__class__.__name__}: {__name__}')
+
+
+def calculate_total_drilling_lengths_m(Configuration, numnonverticalsections: int, nonvertical_length_km: float,
+                                       InputDepth_km: float, OutputDepth_km: float, nprod:int, ninj:int) -> tuple:
+    """
+    returns the total length, vertical length, and non-vertical lengths, depending on the configuration
+    :param Configuration: Configuration of the well
+    :type Configuration: :class:`~geophires
+    :param numnonverticalsections: number of non-vertical sections
+    :type numnonverticalsections: int
+    :param nonvertical_length_km: length of non-vertical sections in km
+    :type nonvertical_length_km: float
+    :param InputDepth_km: depth of the well in km
+    :type InputDepth_km: float
+    :param OutputDepth_km: depth of the output end of the well in km, if U shaped, and not horizontal
+    :type OutputDepth_km: float
+    :param nprod: number of production wells
+    :type nprod: int
+    :param ninj: number of injection wells
+    :return: total length, vertical length, and horizontal lengths in meters
+    :rtype: tuple
+    """
+    if Configuration == Configuration.ULOOP:
+        # Total drilling depth of both wells and laterals in U-loop [m]
+        vertical_pipe_length_m = (nprod * InputDepth_km * 1000.0) + (ninj * OutputDepth_km * 1000.0)
+        nonvertical_pipe_length_m = numnonverticalsections * nonvertical_length_km * 1000.0
+    elif Configuration == Configuration.COAXIAL:
+        # Total drilling depth of well and lateral in co-axial case [m]
+        vertical_pipe_length_m = (nprod + ninj) * InputDepth_km * 1000.0
+        nonvertical_pipe_length_m = numnonverticalsections * nonvertical_length_km * 1000.0
+    elif Configuration == Configuration.VERTICAL:
+        # Total drilling depth of well in vertical case [m]
+        vertical_pipe_length_m = (nprod + ninj) * InputDepth_km * 1000.0
+        nonvertical_pipe_length_m = 0.0
+    elif Configuration == Configuration.L:
+        # Total drilling depth of well in L case [m]
+        vertical_pipe_length_m = (nprod + ninj) * InputDepth_km * 1000.0
+        nonvertical_pipe_length_m = numnonverticalsections * nonvertical_length_km * 1000.0
+    else:
+        raise ValueError(f'Invalid Configuration: {Configuration}')
+
+    tot_pipe_length_m = vertical_pipe_length_m + nonvertical_pipe_length_m
+    return tot_pipe_length_m, vertical_pipe_length_m, nonvertical_pipe_length_m
